feat: move 6 FeedFilter subclasses to commons for KMP
Moved to commons/ui/feeds:
- ChannelFeedFilter
- BookmarkPublicFeedFilter
- BookmarkPrivateFeedFilter
- OldBookmarkPublicFeedFilter
- OldBookmarkPrivateFeedFilter
- ChatroomListNewFeedFilter

IAccount additions:
- bookmarkState: BookmarkListState
- oldBookmarkState: OldBookmarkListState
- isAllHidden(Set<HexKey>): Boolean

Also moved Iterable.replace() utility to commons.

Part of KMP iOS migration tracked in #2238.
