+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/chats/publicChannels/dal/ChannelFeedFilter.kt
@@ -20,29 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.screen.loggedIn.chats.publicChannels.dal
 
-import com.vitorpamplona.amethyst.commons.model.Channel
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.ChangesFlowFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
-
-class ChannelFeedFilter(
-    val channel: Channel,
-    val account: Account,
-) : AdditiveFeedFilter<Note>(),
-    ChangesFlowFilter<Note> {
-    override fun feedKey() = channel
-
-    override fun changesFlow() = channel.changesFlow()
-
-    // returns the last Note of each user.
-    override fun feed(): List<Note> = sort(channel.notes.filterIntoSet { key, it -> account.isAcceptable(it) })
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> =
-        newItems
-            .filter { channel.notes.containsKey(it.idHex) && account.isAcceptable(it) }
-            .toSet()
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+// Re-export from commons for backwards compatibility
+typealias ChannelFeedFilter = com.vitorpamplona.amethyst.commons.ui.feeds.ChannelFeedFilter
